Move Google recaptcha validation to utils. Fix Bug when analysis does not exists in mongodb
